Update tests for new constants locations
- Add tests/mnemosyne/unit/test_constants.py for Mnemosyne constants
- Add tests/psyche/unit/test_constants.py for Psyche constants
- Update test_settings.py in both packages to import from new locations
- Remove tests/shared/test_shared_constants.py (no longer needed)
